Add update and delete endpoints

+++ src/main/java/com/flashcard_shreedev/flashcard_backend/repository/FlashcardRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.flashcard_shreedev.flashcard_backend.entity.Flashcard;
 
-public interface FlashcardRepository extends JpaRepository<Flashcard, Long> {
+public interface FlashcardRepository extends JpaRepository<Flashcard, Integer> {
     //Custom query methods can be defined here if needed
     
 }
